fix: handle double and triple clicks
There were some timing issues causing the tooltip not to appear on
double and triple clicks. Wait 500ms between each click before
rendering the popup and only render once for double and triple clicks.

This does add a 500ms delay for single click selections, since the
event is on mouseup not mousedown. But it seems intentional?

For now it's a feature not a bug :p